:octocat: PHP 7.1, PhpUnit 7.5, proper PSR-4 autoloading
