+++ .travis.yml
@@ -2,12 +2,9 @@ sudo: false
 language: php
 
 php:
-  - 5.4
-  - 5.5
-  - 5.6
-  - 7.0
   - 7.1
   - 7.2
+  - 7.3
   - hhvm
 
 script:
