+++ composer.json
@@ -21,10 +21,10 @@
     }
   },
   "require": {
-    "php": ">=5.4.0"
+    "php": ">=7.1.0"
   },
   "require-dev": {
-    "phpunit/phpunit": "^4.8.35|~5.7"
+    "phpunit/phpunit": "^7.5"
   },
   "extra": {
     "branch-alias": {
