+++ test/RouteCollectorTest.php
@@ -2,7 +2,6 @@
 
 namespace FastRoute\Test;
 
-use FastRoute\RouteCollector;
 use PHPUnit\Framework\TestCase;
 
 class RouteCollectorTest extends TestCase
@@ -101,17 +100,3 @@ class RouteCollectorTest extends TestCase
     }
 }
 
-class DummyRouteCollector extends RouteCollector
-{
-    public $routes = [];
-
-    public function __construct()
-    {
-    }
-
-    public function addRoute($method, $route, $handler)
-    {
-        $route = $this->currentGroupPrefix . $route;
-        $this->routes[] = [$method, $route, $handler];
-    }
-}
